Refactored middleware chain creation and usage
The creation of the middleware is now dynamically done based on the files, reducing the amount of places needed to be updated to introduce a new module (and sets up for future work).

Now, for each request, a middleware chain is created based on the items that apply. This will make it easier to debug to see which items were applied for a specific request (once we have debug logging).